Rescue: bind-libs is in fact just a collection of libraries, that are automatically required by dependencies.
With bind 9.10.1-P1, the shared library packaging policy is being applied to bind, resulting in the collection package bind-libs no longer to exist.

As with all other Shared Libraries, those are not generally listed here, as they tend to change very often (ABI bumps); so we do the same and rely on RPM dependencies to bring us bind-related libs.