removed double documentation page for README
removed additional readme file and made sure that readme in root
folder points to both the source docs and the rendered pages on
readthedocs.io

NOTE: the reason why doc pages weren't added manually in the mkdocs.yml
file is because then the default behaviour of including all the doc
files is lost and each need to be declared explicitly, one by one. This
solution doesn't require the contributors to edit two files when adding
a new documentation.

tl;dr: it's more maintainable without single
entries in the .yml file