hdimage: Use at least FAT16
When creating a FAT container for the EFI partition, we are quite likely
to end up with a FAT12 partition. Unfortunately not all firmware is happy
to read from FAT12 (namely the Raspberry Pi), so we're better off forcing
it to FAT16 or FAT32.

The easiest way to do that is to tell mkfs.vfat to use a single sector per
cluster. That way FAT12 can only be represented by tiny images that we'll
never get.

Signed-off-by: Alexander Graf <agraf@suse.de>