don't use 'iterm' terminfo (fate#325746)
fbiterm sets TERM to 'iterm' but this leads to display problems in the
ncurses-UI of yast when using an Unicode (iso10646) encoded font as basic
text font.

Not sure why this happens as fbiterm otherwise displays Unicode chars just
fine even in this situation. Looks like 'iterm' does not match fbiterm's
capabilities. Swapping it for 'screen' seems to make it work fine.