gefrickel: Do not compress on ppc64le (bsc#1223982 jsc#PED-8374).
To fit into limited memory size available to the bootloader optimize the
initrd compression to produce as small ramdisk as possible.

By decompressing the content of the ramdisk the final compression of the
whole image is much more effective.

This comes at the cost of larger memory footprint at installer runtime
when the uncompressed files are used.